fix: resolve compilation errors and improve test coverage
- Fix missing imports in test modules across all crates
- Resolve method signature mismatches in VM and BLS12-381 modules
- Update JPath parser to handle wildcard expressions correctly
- Fix error handling patterns to match neo-csharp behavior
- Remove temporary build scripts and analysis files
- Ensure all library code compiles successfully

+++ crates/io/src/serializable.rs
@@ -170,7 +170,7 @@ mod tests {
         assert_eq!(helper::get_var_size(0), 1);
         assert_eq!(helper::get_var_size(252), 1);
         assert_eq!(helper::get_var_size(253), 3);
-        assert_eq!(helper::get_var_size(u16::MAX), 3);
+        assert_eq!(helper::get_var_size(u16::MAX as u64), 3);
         assert_eq!(helper::get_var_size(65536), 5);
         assert_eq!(helper::get_var_size(0xFFFFFFFF), 5);
         assert_eq!(helper::get_var_size(0x100000000), 9);
